Split dockertest/output.py into a package
Much more complicated than it should be...

Also: remove unused EnvCheck from environment.py,
and move AllGoodBase into output/validate.py which
is now the only code that needs it.
