--- dockertest/output.py
@@ -1,988 +1,0 @@
-"""
-Handlers for command-line output processing, crash/panic detection, etc.
-"""
-
-# Pylint runs from a different directory, it's fine to import this way
-# pylint: disable=W0403
-
-import re
-import datetime
-from collections import Mapping, MutableSet, Sequence
-from autotest.client import utils
-import subprocess
-import xceptions
-from environment import AllGoodBase
-from version import LooseVersion
-
-
-class DockerVersion(object):
-
-    """
-    Parser of docker-cli version command output as client/server properties
-
-    :param version_string: Raw, possibly empty or multi-line output
-                           from docker version command.
-    """
-    #: Raw, possibly empty or multi-line output from docker version command.
-    #: Read-only, set in __init__
-    version_string = None
-    #: Various cache's for properties (do not use)
-    _client = None
-    _server = None
-    _client_lines = None
-    _server_lines = None
-    _client_info = None
-    _server_info = None
-    _has_distinct_exit_codes = None
-
-    def __init__(self, version_string=None, docker_path=None):
-        # If called without an explicit version string, run docker to find out
-        if version_string is None:
-            if docker_path is None:
-                docker_path = 'docker'
-            version_string = subprocess.check_output(docker_path + ' version',
-                                                     shell=True,
-                                                     close_fds=True)
-        self.version_string = version_string
-
-    def _oops(self, what):
-        raise xceptions.DockerOutputError("Couldn't parse %s "
-                                          "from %s" % (what,
-                                                       self.version_string))
-
-    # This is old code, not updating to preserve old behavior
-    def _old_client(self):
-        if self._client is None:
-            regex = re.compile(r'Client\s+version:\s+(\d+\.\d+\.\d+\S*)',
-                               re.IGNORECASE)
-            mobj = None
-            for line in self.version_lines:
-                mobj = regex.search(line.strip())
-                if bool(mobj):
-                    self._client = mobj.group(1)
-        if self._client is None:
-            self._oops('client version')
-        return self._client
-
-    # This is old code, not updating to preserve old behavior
-    def _old_server(self):
-        if self._server is None:
-            regex = re.compile(r'Server\s*version:\s*(\d+\.\d+\.\d+\S*)',
-                               re.IGNORECASE)
-            mobj = None
-            for line in self.version_lines:
-                mobj = regex.search(line.strip())
-                if bool(mobj):
-                    self._server = mobj.group(1)
-        if self._server is None:
-            self._oops('server version')
-        return self._server
-
-    def _split_client_server(self):
-        # Split the raw string into client & server sections
-        client_lines = []
-        server_lines = []
-        version_lines = list(self.version_lines)  # work on a copy
-        version_lines.reverse()  # start at beginning
-        while version_lines:
-            version_line = version_lines.pop().strip()
-            if version_line == '':
-                continue
-            elif version_line.find('Client:') > -1:
-                # Don't assume which section came first
-                while version_lines and version_lines[-1].find('Server:') < 0:
-                    version_line = version_lines.pop().strip()
-                    if version_line == '':
-                        continue
-                    client_lines.append(version_line)
-                continue
-            elif version_line.find('Server:') > -1:
-                # Don't assume which section came first
-                while version_lines and version_lines[-1].find('Client:') < 0:
-                    version_line = version_lines.pop().strip()
-                    if version_line == '':
-                        continue
-                    server_lines.append(version_line)
-                continue
-            else:
-                msg = ("Unexpected line '%s' in version string: '%s'"
-                       % (version_line, self.version_string))
-                raise xceptions.DockerOutputError(msg)
-        return (client_lines, server_lines)
-
-    # This is to preserve API behavior for old tests
-    @property
-    def version_lines(self):
-        """Read-only property that returns all lines in version table"""
-        return self.version_string.splitlines()
-
-    @property
-    def client_lines(self):
-        """
-        Read-only property of split/stripped client section of version string
-        """
-        if not self._client_lines:
-            (self._client_lines,
-             self._server_lines) = self._split_client_server()
-        return self._client_lines
-
-    @property
-    def server_lines(self):
-        """
-        Read-only property of split/stripped server section of version string
-        """
-        if not self._server_lines:
-            (self._client_lines,
-             self._server_lines) = self._split_client_server()
-        return self._server_lines
-
-    def _info(self, is_client, key):
-        key = key.strip()
-        if is_client:
-            infodict = self._client_info
-            infolines = self.client_lines
-        else:
-            infodict = self._server_info
-            infolines = self.server_lines
-        try:
-            return infodict[key.strip().lower()]
-        except TypeError:  # infodict == None
-            if is_client:
-                self._client_info = {}
-            else:
-                self._server_info = {}
-            return self._info(is_client, key)
-        except KeyError:  # infodict == empty
-            if not infodict:
-                for line in infolines:
-                    try:
-                        _key, value = line.strip().lower().split(':', 1)
-                    except ValueError:
-                        raise ValueError("Error splitting info line '%s'"
-                                         % line)
-                    infodict[_key.strip()] = value.strip()
-                return self._info(is_client, key)
-            else:  # avoid infinite recursion
-                self._oops("key %s" % key)
-
-    def client_info(self, key):
-        """Return item named 'key' from client section of version info table"""
-        return self._info(True, key)
-
-    def server_info(self, key):
-        """Return item named 'key' from server section of version info table"""
-        return self._info(False, key)
-
-    @property
-    def client(self):
-        """
-        Read-only property representing version-number string of docker client
-        """
-        if self._client is None:
-            try:
-                self._client = self._old_client()
-            except xceptions.DockerOutputError:
-                self._client = self.client_info('version')
-        if self._client is None:
-            self._oops('client version')
-        return self._client
-
-    @property
-    def server(self):
-        """
-        Read-only property representing version-number string of docker server
-        """
-        if self._server is None:
-            try:
-                self._server = self._old_server()
-            except xceptions.DockerOutputError:
-                self._server = self.server_info('version')
-        if self._server is None:
-            self._oops('server version')
-        return self._server
-
-    @staticmethod
-    def _require(wanted, name, other_version):
-        required_version = LooseVersion(wanted)
-        if other_version < required_version:
-            msg = ("Test requires docker %s version >= %s; %s found"
-                   % (name, required_version, other_version))
-            raise xceptions.DockerTestNAError(msg)
-        # In case it's useful to caller
-        return other_version
-
-    def require_server(self, wanted):
-        """
-        Run 'docker version', parse server version, compare to wanted.
-
-        :param wanted: required docker (possibly remote) server version
-        :raises DockerTestNAError: installed docker < wanted
-        """
-        return self._require(wanted, 'server', self.server)
-
-    def require_client(self, wanted):
-        """
-        Run 'docker version', parse client version, compare to wanted.
-
-        :param wanted: required docker client version
-        :raises DockerTestNAError: installed docker < wanted
-        """
-        return self._require(wanted, 'client', self.client)
-
-    @property
-    def has_distinct_exit_codes(self):
-        """
-        2016-03-23 **TEMPORARY** - for transition from docker-1.9 to 1.10
-
-        docker-1.10 will introduce distinct exit codes to allow differentiating
-        between container status and errors from docker run itself; see
-        bz1097344 and docker PR14012. If we see an exit code of 125 here,
-        assume we're using the new docker.
-        """
-        if self._has_distinct_exit_codes is None:
-            try:
-                # docker-1.10 *must* support distinct exit codes
-                self.require_client('1.10')
-                has = True
-            except xceptions.DockerTestNAError:
-                # some builds of 1.9 might support it. (FIXME: really?)
-                d_run = utils.run('docker run --invalid-opt invalid-image',
-                                  ignore_status=True)
-                has = (d_run.exit_status > 120)
-            DockerVersion._has_distinct_exit_codes = has
-        return DockerVersion._has_distinct_exit_codes
-
-
-class ColumnRanges(Mapping):
-
-    """
-    Immutable map of start/end offsets to/from column names.
-
-    :param header: Table header string of multi-space separated column names
-    :param expected: Precise number of columns expected, or raise ValueError
-    :param min_col_len: Minimum number of characters for a column header
-    :raises ValueError: Column < than min_col_len or # columns != expected
-    """
-
-    # Too few pub. methods, pylint doesn't count abstract __special_methods__
-    # pylint: disable=R0903
-
-    __slots__ = ('ranges', 'columns', 'count')
-
-    #: Iterable of start/end character-offset tuples corresponding to columns
-    ranges = None
-
-    #: Iterable of column names corresponding to ranges
-    columns = None
-
-    #: Number of columns/ranges
-    count = None
-
-    #: Regex specifying the column separator
-    _re = re.compile(r"\s\s+")
-
-    def __init__(self, header, min_col_len=3, expected=None):
-        header_strip = header.strip()  # just in case
-        cols = [col for col in self._re.split(header_strip)]
-        if expected is not None and len(cols) != expected:
-            raise ValueError("Columns parsed (%s) does not match expected (%s)"
-                             % (cols, expected))
-        columns = []  # converted to set at end then discarded
-        starts = []  # zip()'d at end then discarded
-        for col in cols:
-            col_strip = col.strip()
-            if len(col_strip) < min_col_len:
-                raise ValueError("Column name '%s' is smaller than minimum "
-                                 "required column name length %d"
-                                 % (col_strip, min_col_len))
-            columns.append(col_strip)
-            starts.append(header_strip.index(col_strip))  # lookup char offset
-        ends = starts[1:] + [None]  # ending offsets EXCLUSIVE for range()
-        # Stored separetly b/c dict() storage would be un-ordered!
-        self.columns = tuple(columns)
-        ranges = zip(starts, ends)  # needed for exception message
-        self.ranges = tuple(ranges)
-        self.count = len(columns)  # allow check for duplicates vs set()
-        # Check duplicate column names or ranges
-        if (self.count != len(set(self.ranges)) or
-                self.count != len(set(self.columns))):
-            raise ValueError("Duplicate column names '%s' or ranges '%s' "
-                             "detected: " % (columns, ranges))
-
-    def __str__(self):
-        lst = [("%s: %s-%s" % (col, start, end))
-               for col, (start, end) in zip(self.columns, self.ranges)
-               if end > 0]
-        lst.append("%s: %s-End" % (self.columns[-1], self.ranges[-1][0]))
-        return ", ".join(lst)
-
-    def __repr__(self):
-        return "<%s object of {%s}>" % (self.__class__.__name__,
-                                        str(self))
-
-    def __len__(self):
-        return self.count  # instance is immutable
-
-    def __contains__(self, item):
-        return item in self.ranges or item in self.columns
-
-    def __iter__(self):
-        return self.ranges.__iter__()
-
-    def __getitem__(self, key):
-        try:
-            return self.columns[list(self.ranges).index(key)]
-        except ValueError:
-            return self.ranges[list(self.columns).index(key)]
-
-    def offset(self, offset):
-        """
-        Return column name corresponding to range containing offset
-
-        :param offset: Character offset number to lookup
-        :returns: Column name string
-        :raises IndexError: if offset not found in any header range
-        """
-        if offset is None or offset < 0:
-            return self.columns[-1]
-        for index, (start, end) in enumerate(self.ranges):
-            if end is None:
-                return self.columns[-1]
-            if offset in xrange(start, end):
-                return self.columns[index]
-        if offset > 0:
-            return self.columns[-1]  # beyond end of any ranges
-
-
-class TextTable(MutableSet, Sequence):
-
-    """
-    Parser for tabular data with values separated by character offsets
-
-    :param table: String of table header, optionally followed by data rows
-    :param columnranges: Optional ColumnRanges instance to use
-    :param header: Optional header to use, instead of first-line,
-                   ignored if columnranges parameter is non-None
-    :param tabledata: Optional lines of data to use, optionally with header
-                      option or columnranges option (but not both).
-    :raises TypeError: if table contains less than one line
-    :raises ValueError: if key_column is not in table_columns
-    """
-
-    #: Permit duplicate rows to be added
-    allow_duplicate = False
-
-    #: Comparison function to use when sorting
-    compare = None
-
-    # internal cache of column name to tuple of start,end offset range
-    columnranges = None
-
-    #: internal cache of parsed rows
-    _rows = None
-
-    def __init__(self, table, columnranges=None, header=None, tabledata=None):
-        if columnranges is not None and header is not None:
-            raise ValueError("Cannot specify both columnranges and header "
-                             "parameters")
-
-        if header is None and tabledata is None:
-            table_lines = table.strip().splitlines()
-            if len(table_lines) < 1:
-                # FIXME: This should probably be a ValueError
-                raise TypeError("Table shorter than one line: %s" % table)
-            header, tabledata = self.parseheader(table)
-        elif header is None:
-            # tabledata == tabledata
-            table_lines = tabledata.strip().splitlines()
-            header = table_lines[0]
-        else:  # tabledata == none
-            tabledata = table_lines.strip().splitlines()
-
-        if columnranges is None:
-            # First line is header
-            self.columnranges = ColumnRanges(header)
-        else:
-            if not isinstance(columnranges, ColumnRanges):
-                raise TypeError("columnranges is not a ColumnRanges instance")
-            self.columnranges = columnranges
-
-        self._rows = []
-
-        if tabledata is not None:
-            for line in self.parserows(tabledata):
-                line_strip = line.strip()
-                self.append(self.parse_line(line_strip))
-
-    def __eq__(self, other):
-        if not hasattr(other, '__iter__'):
-            return False
-        return list(self._rows) == list(other)
-
-    def __len__(self):
-        return len(self._rows)
-
-    def __iter__(self):
-        return self._rows.__iter__()
-
-    def __contains__(self, value):
-        """
-        Return true if any row or row[self.key_column] equals value
-        """
-        return self._rows.__contains__(value)
-
-    def __setitem__(self, index, value):
-        self.conform_or_raise(value)
-        return self._rows.__setitem__(index, value)
-
-    def __delitem__(self, index):
-        return self._rows.__delitem__(index)
-
-    def __getitem__(self, index):
-        return self._rows.__getitem__(index)
-
-    def insert(self, index, value):
-        """
-        Insert value contents at index
-        """
-        self.conform_or_raise(value)
-        return self._rows.insert(index, value)
-
-    def add(self, value):
-        self.conform_or_raise(value)
-        return self._rows.append(value)
-
-    def discard(self, index):
-        """
-        Wraps del texttable[index[
-        """
-        return self.__delitem__(index)
-
-    def append(self, value):
-        """
-        Inserts value item or iterable at end
-        """
-        self.conform_or_raise(value)
-        return self._rows.append(value)
-
-    def conforms(self, value):
-        """
-        Return True if value is non-duplicate dict-like with all column keys
-        """
-        try:
-            self.conform_or_raise(value)
-            return True
-        except ValueError:
-            return False
-
-    def conform_or_raise(self, value):
-        """Raise ValueError if not self.conforms(value)"""
-        if not isinstance(value, dict):
-            raise ValueError("Value '%s' is not a dict-like" % value)
-        keys = set(value.keys())
-        expected = set(self.columnranges.values())
-        if keys == expected:
-            if not self.allow_duplicate and self.__contains__(value):
-                raise ValueError("Value '%s' is duplicate" % value)
-        else:
-            raise ValueError("Value's keys %s != %s columns"
-                             % (keys, expected))
-
-    @staticmethod
-    def value_filter(value):
-        """
-        Converts empty stripped or '<none>' strings into None value
-        """
-        value = str(value).strip()
-        if value == '':
-            return None
-        if value == '<none>':
-            return None
-        return value
-
-    @staticmethod
-    def parseheader(table):
-        """
-        Parse string into tuple(header, data)
-        """
-        lines = table.strip().splitlines()
-        if len(lines) < 1:
-            raise TypeError("Table shorter than one line: %s" % table)
-        if len(lines) == 1:
-            tabledata = None
-        else:
-            tabledata = "\n".join(lines[1:])  # put back together
-        return (lines[0], tabledata)  # both regular strings
-
-    @staticmethod
-    def parserows(tabledata):
-        """
-        Parse table data string (minus header line) into a list of rows
-        """
-        return tabledata.strip().splitlines()
-
-    def parse_line(self, line):
-        """
-        Parse one line into a dict based on columnranges
-        """
-        newdict = {}
-        strippedline = line.strip()
-        for (start, end), colname in self.columnranges.items():
-            newdict[colname] = self.value_filter(strippedline[start:end])
-        return newdict
-
-    def search(self, col_name, value, match_func=None):
-        """
-        Returns a list of dictionaries containing col_name key with value
-
-        :param col_name: Column name string to use
-        :param value: Value to compare each row's column name to
-        :match_func: If specified, match found when
-                     match_func(col_name, value, row_value) returns True
-        """
-        result = []
-        for row in self._rows:
-            if match_func is None:
-                if row.get(col_name) == value:
-                    result.append(dict(row))
-            else:
-                if match_func(col_name, value, row.get(col_name)):
-                    result.append(dict(row))
-        return result
-
-    def find(self, col_name, value, match_func=None):
-        """
-        Return dictionary with key col_name == value raise IndexError if != 1
-        """
-        found = self.search(col_name, value, match_func)
-        if len(found) != 1:
-            raise IndexError("Found %d rows with %s == %s"
-                             % (len(found), col_name, value))
-        return found[0]
-
-
-class OutputGoodBase(AllGoodBase):
-
-    """
-    Compare True if all methods ending in '_check' return True on stdout/stderr
-
-    :param cmdresult: autotest.client.utils.CmdResult instance
-    :param ignore_error: Raise exceptions.DockerOutputError if False
-    :param skip: Iterable of checks to bypass, None to run all
-    """
-
-    #: Reference to original CmdResult instance
-    cmdresult = None
-
-    #: Stripped standard-output string
-    stdout_strip = None
-
-    #: Stripped standard-error string
-    stderr_strip = None
-
-    def __init__(self, cmdresult, ignore_error=False, skip=None):
-        # Base class __init__ is abstract
-        # pylint: disable=W0231
-        self.cmdresult = cmdresult
-        self.stdout_strip = cmdresult.stdout.strip()
-        self.stderr_strip = cmdresult.stderr.strip()
-        # All methods called twice with mangled names, mangle skips also
-        if skip is not None:
-            if isinstance(skip, (str, unicode)):
-                skip = [skip]
-            newskip = []
-            for checker in skip:
-                newskip.append(checker + '_stdout')
-                newskip.append(checker + '_stderr')
-        else:
-            newskip = skip
-        self.__instattrs__(newskip)
-        for checker in [name for name in dir(self) if name.endswith('_check')]:
-            self.callables[checker + '_stdout'] = getattr(self, checker)
-            self.callables[checker + '_stderr'] = getattr(self, checker)
-        self.call_callables()
-        # Not nonzero means One or more checkers returned False
-        if not ignore_error and not self.__nonzero__():
-            # Str representation will provide details
-            raise xceptions.DockerOutputError(str(self))
-
-    def __str__(self):
-        if not self.__nonzero__():
-            msg = super(OutputGoodBase, self).__str__()
-            return "%s\nSTDOUT:\n%s\nSTDERR:\n%s" % (msg, self.stdout_strip,
-                                                     self.stderr_strip)
-        else:
-            return super(OutputGoodBase, self).__str__()
-
-    def callable_args(self, name):
-        if name.endswith('_stdout'):
-            return {'output': self.stdout_strip}
-        elif name.endswith('_stderr'):
-            return {'output': self.stderr_strip}
-        else:
-            raise RuntimeError("Unexpected check method name %s" % name)
-
-    def prepare_results(self, results):
-        duplicate = False
-        for checker, passed in results.items():
-            if not passed and not duplicate:
-                exit_status = self.cmdresult.exit_status
-                stdout = self.cmdresult.stdout.strip()
-                stderr = self.cmdresult.stderr.strip()
-                detail = 'Command '
-                if exit_status != 0:
-                    detail += 'exit %d ' % exit_status
-                if len(stdout) > 0:
-                    detail += 'stdout "%s" ' % stdout
-                if len(stderr) > 0:
-                    detail += 'stderr "%s".' % stderr
-                self.details[checker] = detail
-                duplicate = True  # all other failures will be same
-        return super(OutputGoodBase, self).prepare_results(results)
-
-
-class OutputGood(OutputGoodBase):
-
-    """
-    Container of standard checks
-    """
-
-    @staticmethod
-    def crash_check(output):
-        """
-        Return False if Go panic string found in output
-
-        :param output: Stripped output string
-        :return: True if Go panic pattern **not** found
-        """
-        regex = re.compile(r'\s*panic:\s*.+error.*')
-        for line in output.splitlines():
-            if bool(regex.search(line.strip())):
-                return False  # panic message found
-        return True  # panic message not found
-
-    @staticmethod
-    def usage_check(output):
-        """
-        Return False if 'Docker usage' pattern found in output
-
-        :param output: Stripped output string
-        :return: True if usage message pattern **not** found
-        """
-        regex = re.compile(r'\s*usage:\s+docker\s+.*', re.IGNORECASE)
-        for line in output.splitlines():
-            if bool(regex.search(line.strip())):
-                return False  # usage message found
-        return True  # usage message not found
-
-    @staticmethod
-    def error_check(output):
-        """
-        Return False if 'Error: ' pattern found in output
-
-        :param output: Stripped output string
-        :return: True if 'Error: ' does **not** sppear
-        """
-        for line in output.splitlines():
-            return line.lower().strip().find('error') == -1
-        return True
-
-    @staticmethod
-    def fata_check(output):
-        """
-        Return False if 'FATA[xxxx]' pattern found in output
-
-        :param output: Stripped output string
-        :return: True if 'FATA ' does **not** sppear
-        """
-        regex = re.compile(r'FATA\[\d+')
-        return not bool(regex.search(output))
-
-
-class OutputNotBad(OutputGood):
-
-    """
-    Same as OutputGood, except skip checking error/usage messages by default
-    """
-
-    #: Full command line that outputs string to search for kernel oopses
-    GET_PANIC_CMD = ('journalctl --no-pager --all --reverse '
-                     '--dmesg --boot --priority=warning')
-
-    #: Kernel oops string to look for
-    OOPS_STRING = 'oops'
-
-    #: Caches output value from running GET_PANIC_CMD, None if Error
-    _dmesg_cache = None
-
-    def __init__(self, cmdresult, ignore_error=False, skip=None):
-        defaults = ['error_check', 'usage_check']
-        if skip is None:
-            skip = defaults
-        else:
-            if isinstance(skip, basestring):
-                skip = defaults + [skip]
-            else:
-                skip = defaults + skip
-        super(OutputNotBad, self).__init__(cmdresult, ignore_error, skip)
-
-    def kernel_panic(self, output):
-        """
-        Checks output from ``GET_PANIC_CMD`` for ``OOPS_STRING``
-        """
-        del output  # not used
-        return self.dmesg.lower().strip().find(self.OOPS_STRING) == -1
-
-    @property
-    def dmesg(self):
-        """
-        Represents (cached) ``GET_PANIC_CMD`` output last obtained.
-        """
-        if self._dmesg_cache is None:
-            self._dmesg_cache = subprocess.check_output(self.GET_PANIC_CMD,
-                                                        shell=True)
-        return self._dmesg_cache
-
-
-def wait_for_output(output_fn, pattern, timeout=60, timestep=0.2):
-    r"""
-    Wait for matched_string in async_process.stdout max for time==timeout.
-
-    :param process_output_fn: function which returns data for matching.
-    :type process_output_fn: function
-    :param pattern: string which should be found in stdout.
-    :return: True if pattern matches process_output else False
-    """
-    if not callable(output_fn):
-        raise TypeError("Output function type %s value %s is not a callable"
-                        % (output_fn.__class__.__name__, str(output_fn)))
-    regex = re.compile(pattern)
-    _fn = lambda: regex.findall(output_fn()) != []
-    res = utils.wait_for(_fn, timeout, step=timestep)
-    if res:
-        return True
-    return False
-
-
-def mustpass(cmdresult, failmsg=None):
-    """
-    Check docker cmd results for pass. Raise exception when command failed.
-
-    :param cmdresult: results of cmd.
-    :type cmdresult: object convertible to string with variable exit_status
-    :param failmsg: Additional messages for describing problem when cmd fails.
-    """
-    if failmsg is None:
-        details = "%s" % cmdresult
-    else:
-        details = "%s\n%s" % (failmsg, cmdresult)
-    OutputNotBad(cmdresult)
-    if cmdresult.exit_status != 0:
-        raise xceptions.DockerExecError("Unexpected non-zero exit code,"
-                                        " details: %s" % details)
-    return cmdresult
-
-
-def mustfail(cmdresult, expected_status=None, failmsg=None):
-    """
-    Check docker cmd results for pass. Raise exception when command passed.
-
-    :param cmdresult: results of cmd.
-    :type cmdresult: object convertible to string with variable exit_status
-    :param expected_status: exit status we expect to see from cmd.
-                            NOTE: for backward compatibility with pre-20160331
-                            code, we try to gracefully handle if this param
-                            is missing (default to 1) or a string (set failmsg)
-    :type expected_status: integer, 1-255
-    :param failmsg: Additional messages for describing problem when cmd fails.
-    """
-    # FIXME: temporary: backward compatibility for pre-20160330 code
-    # FIXME: remove before the next API-changing release
-    if expected_status is None:                    # old: mustfail(x)
-        expected_status = 1
-    if isinstance(expected_status, basestring):    # old: mustfail(x, "msg")
-        if not expected_status.isdigit():     # pylint: disable=E1101
-            failmsg = expected_status
-            expected_status = 1
-
-    if failmsg is None:
-        details = "%s" % cmdresult
-    else:
-        details = "%s\n%s" % (failmsg, cmdresult)
-    OutputNotBad(cmdresult)
-    if cmdresult.exit_status == expected_status:
-        return cmdresult
-    # On pre-1.10 docker, accept any nonzero exit status: it's impossible
-    # to automatically map docker-1.10 codes to 1.9
-    # FIXME: temporary; remove once we no longer run on pre-1.10 docker
-    if not DockerVersion().has_distinct_exit_codes:
-        if cmdresult.exit_status != 0:
-            return cmdresult
-
-    raise xceptions.DockerExecError("Unexpected exit code %d; expected %d."
-                                    " Details: %s" % (
-                                        cmdresult.exit_status,
-                                        expected_status,
-                                        details))
-
-
-# This class inherits a LOT of public methods, but most of what
-# appears here is all 'behind the scenes' stuff for producing
-# immutable instances.
-class DockerTime(datetime.datetime):  # pylint: disable=R0903
-
-    """
-    Immutable Docker specialized zulu-time representation
-
-    :note: value is undefined when instance == instance.tzinfo.EPOCH.
-           For example, the FinishedAt time of a still-running container
-    :param isostr: ISO 8601 format string
-    :param sep: Optional separation character ('T' by default)
-    :raise: ValueError if isostr is unparseable
-    """
-
-    class UTC(datetime.tzinfo):
-
-        """Singleton representation of UTC timezone and epoch point"""
-
-        #: Zero timedelta offset
-        ZERO = datetime.timedelta(0)
-
-        #: Representation of the beginning of time, created in __new__()
-        EPOCH = datetime.datetime(year=1, month=1, day=1,
-                                  hour=0, minute=0, second=0,
-                                  microsecond=0)
-
-        #: Reference to the singleton instance
-        singleton = None
-
-        def __new__(cls):
-            if cls.singleton is None:
-                cls.singleton = super(DockerTime.UTC, cls).__new__(cls)
-                cls.EPOCH = cls.EPOCH.replace(tzinfo=cls.singleton)
-                # Not an invalid attribute name, this is a constant
-                # value that requires one update to reach it's final
-                # form.
-                cls.singleton.EPOCH = cls.EPOCH  # pylint: disable=C0103
-            return cls.singleton
-
-        @classmethod
-        def utcoffset(cls, dt):
-            del dt  # not needed
-            return cls.ZERO
-
-        @classmethod
-        def tzname(cls, dt):
-            del dt  # not needed
-            return "UTC"
-
-        @classmethod
-        def dst(cls, dt):
-            del dt  # not needed
-            return cls.ZERO
-
-    class UTCOffset(datetime.tzinfo):
-
-        """Fixed offset in hours and minutes from UTC."""
-
-        def __init__(self, offset_string):
-            super(DockerTime.UTCOffset, self).__init__()
-            numbers = offset_string.split(':')
-            hours = int(numbers[0])
-            minutes = int(numbers[1])
-            self.__offset = datetime.timedelta(hours=hours, minutes=minutes)
-            self.__name = "UTC%s" % offset_string
-
-        def utcoffset(self, dt):
-            del dt  # not used, but specified in base
-            return self.__offset
-
-        def tzname(self, dt):
-            del dt  # not used, but specified in base
-            return self.__name
-
-        def dst(self, dt):
-            del dt  # not used, but specified in base
-            return DockerTime.UTC.ZERO
-
-    def __new__(cls, isostr, sep=None):
-        if sep is None:
-            sep = 'T'
-        # datetime can output zulu time but not consume it.
-        base = "%s%s%s" % (r"(\s*\d{4})-(\d{2})-(\d{2})",
-                           sep,
-                           r"(\d{2}):(\d{2}):(\d{2})")
-        keys = ['year', 'month', 'day',
-                'hour', 'minute', 'second']
-        values = []
-        # Order is significant, some parsers depend on one-another
-        parsers = [cls.__new_tzoffset__, cls.__new_zulu__, cls.__new_us__]
-        for parser in parsers:
-            if parser(isostr, base, values, keys, cls.UTC()):
-                break  # Parsers return True on success
-        if values == []:  # No parser was succesful
-            raise ValueError("Malformed date time string %s" % isostr)
-        # Convert any strings into integers
-        for index, value in enumerate(values):
-            if isinstance(value, basestring):
-                values[index] = int(value)
-        dargs = dict(zip(tuple(keys), tuple(values)))
-        return super(DockerTime, cls).__new__(cls, **dargs)
-
-    def __repr__(self):
-        return '{0}("{1:%Y-%m-%dT%H:%M:%S}.{2:06d}{1:%z}")'.format(
-            self.__class__.__name__, self, self.microsecond)
-
-    @classmethod
-    def __new_zulu__(cls, isostr, base, values, keys, tzn):
-        # Killall letter Z and z's
-        isostr.replace('z', ' ')
-        isostr.replace('Z', ' ')
-        # may or may not have fractional seconds
-        has_us = cls.__new_us__(isostr, base, values, keys, tzn)
-        if has_us:
-            return True
-        else:
-            regex = re.compile(base)
-            mobj = regex.search(isostr)
-            if mobj:
-                values += list(mobj.groups())
-                keys.append('tzinfo')
-                values.append(tzn)
-                return True
-            return False
-
-    @classmethod
-    def __new_us__(cls, isostr, base, values, keys, tzn):
-        # Try with interpreted microseconds
-        regex = re.compile(base + r"\.(\d+)")
-        mobj = regex.search(isostr)
-        if mobj:
-            values += list(mobj.groups())
-            # Convert seconds decimal fraction into microseconds
-            sec_frac = float("0.%s" % values[-1])
-            values[-1] = int(sec_frac * 1000000)
-            keys.append('microsecond')
-            values.append(tzn)
-            keys.append('tzinfo')
-            return True
-        return False
-
-    @classmethod
-    def __new_tzoffset__(cls, isostr, base, values, keys, tzn):
-        # Check if ends with +/-00:00 timezone offset, optional
-        # non-capturing seconds-fraction parsed by __new_us__()
-        regex = re.compile(base + r"(?:\.(\d+))?([+-]{1}\d{2}:\d{2})")
-        mobj = regex.search(isostr)
-        if mobj:
-            tzn = cls.UTCOffset(mobj.group(8))
-            # Remove timezone from string, attempt parsing with __new_us__
-            isostr = isostr[0:len(isostr) - len(mobj.group(8))]
-            return cls.__new_us__(isostr, base, values, keys, tzn)
-        return False
-
-    def is_undefined(self):
-        """
-        Return True if this instance represents an undefined date & time
-        """
-        return self - self.UTC.singleton.EPOCH == self.UTC.ZERO
