+++ run_pylint.sh
@@ -34,6 +34,7 @@ INIT_HOOK="
 AP = os.environ.get('AUTOTEST_PATH', '/usr/local/autotest')
 sys.path.append(os.path.abspath(AP + '/..'))
 sys.path.append(os.path.abspath('.'))
+sys.path.append(os.path.abspath('dockertest'))
 import autotest
 import autotest.common
 "
